Refactor hive pkg list function
