+++ pkg/hive/clusterdeployment.go
@@ -177,37 +177,6 @@ func (builder *ClusterDeploymentBuilder) Get() (*hiveV1.ClusterDeployment, error
 	return clusterDeployment, err
 }
 
-// ListClusterDeploymentsInAllNamespaces returns a cluster-wide clusterdeployment inventory.
-func ListClusterDeploymentsInAllNamespaces(
-	apiClient *clients.Settings,
-	options goclient.ListOption) ([]*ClusterDeploymentBuilder, error) {
-	glog.V(100).Infof("Listing all clusterdeployments with the options %v", options)
-
-	clusterDeployments := new(hiveV1.ClusterDeploymentList)
-	err := apiClient.List(context.TODO(), clusterDeployments, options)
-
-	if err != nil {
-		glog.V(100).Infof("Failed to list all clusterDeployments due to %s", err.Error())
-
-		return nil, err
-	}
-
-	var clusterDeploymentObjects []*ClusterDeploymentBuilder
-
-	for _, clusterDeployment := range clusterDeployments.Items {
-		copiedClusterDeployment := clusterDeployment
-		clusterDeploymentBuilder := &ClusterDeploymentBuilder{
-			apiClient:  apiClient,
-			Object:     &copiedClusterDeployment,
-			Definition: &copiedClusterDeployment,
-		}
-
-		clusterDeploymentObjects = append(clusterDeploymentObjects, clusterDeploymentBuilder)
-	}
-
-	return clusterDeploymentObjects, nil
-}
-
 // PullClusterDeployment pulls existing clusterdeployment from cluster.
 func PullClusterDeployment(apiClient *clients.Settings, name, nsname string) (*ClusterDeploymentBuilder, error) {
 	glog.V(100).Infof("Pulling existing clusterdeployment name %s under namespace %s from cluster", name, nsname)
